pyproject.toml
[tool.poetry]
authors = ["bryan.matteson <bryan.matteson@cbinsights.com>"]
description = ""
name = "cbiproto"
packages = [
  {include = "cbiproto"},
  {include = "cbiproto_tools"},
]
version = "0.0.1"

[tool.poetry.dependencies]
GitPython = {version = "^3.1.27", optional = true}
Jinja2 = {version = "^3.1.2", optional = true}
black = {version = "^19.10b0", optional = true}
cbisyncer = {path = "../cbisyncer", develop = true}
grpc-stubs = {version = "^1.24.10", optional = true}
grpcio-tools = {extras = ["protobuf"], version = "^1.46.0", optional = true}
grpclib = "^0.4.2"
isort = {version = "^5.10.1", optional = true}
mypy-protobuf = {version = "^1.23.0", optional = true}
pydantic = "^1.9.1"
python = "^3.8"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.12.2"
pytest = "^5.2"

[tool.poetry.scripts]
cbiproto-compile = "cbiproto_tools.compile.main:main"
cbiproto-fetch = "cbiproto_tools.fetch.main:main"
protoc-gen-python_cbiproto = "cbiproto.plugin:main"

[tool.poetry.extras]
compiler = ["jinja2", "black", "isort", "GitPython", "grpc-stubs", "grpcio-tools", "mypy-protobuf"]

[tool.poe]
include = ["tasks.toml"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 120
multi_line_output = 3
use_parentheses = true

[tool.pycln]
all = true

[tool.black]
include = "\\.(py)$"
line-length = 120
target-version = ["py38"]

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
show_traceback = true

[tool.pyright]
exclude = ["**/__pycache__"]
ignore = []
include = ["src"]
typeCheckingMode = "basic"

reportMissingImports = true
reportMissingTypeStubs = false

pythonPlatform = "All"
pythonVersion = "3.9"
